Close canvas lifecycle review boundaries
